backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
google-auth
google-api-python-client
requests
numpy
python-dateutil
gunicorn
pydantic-settings
jinja2
# Dependencies for Marta AI Assistant
langchain-google-vertexai
langchain
pytz
google-auth-oauthlib
langchain-community
faiss-cpu
langchain-text-splitters 